Move coding (early stages; not tested). This DOES NOT complie due to linking errors, probably inheritance problems. Suggested reduction to a single Board class.

+++ ScrabbleJunior/ScrabbleJunior.cpp
@@ -7,7 +7,7 @@
 
 int main()
 {
-    srand(time(NULL)); //letter randomize
+    srand((unsigned int) time(NULL)); //letter randomize
 
     //english alphabet 
     std::string englishAlphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
